Clean up routes and test files
